+++ tests/checks/isFunc.Test.js
@@ -1,20 +1,20 @@
-import { isFunc } from '../../src';
+import { isFunction } from '../../src';
 import { multiTest, functionData as data } from '../TestUtil';
 
-describe('isFunc', () => {
+describe('isFunction', () => {
 	multiTest({
 		values: data.true,
-		test: (value) => isFunc(value),
+		test: (value) => isFunction(value),
 		assertion: 'isTrue'
 	});
 	multiTest({
 		values: data.false,
-		test: (value) => isFunc(value),
+		test: (value) => isFunction(value),
 		assertion: 'isFalse'
 	});
 	multiTest({
 		values: data.true,
-		test: (value) => isFunc(value, true),
+		test: (value) => isFunction(value, true),
 		assertion: 'isTrue'
 	});
 });