+++ tests/enforcer/types/enforceFunc.Test.js
@@ -4,12 +4,12 @@ import { validFunctions } from '../../TestUtil';
 import { runNegativeTests } from '../enforceTestUtility';
 
 describe('enforce', () => {
-	describe('.func', () => {
+	describe('.function', () => {
 		it('should return the setter value when an array is provided', () => {
-			assert.deepEqual(enforce.func(validFunctions[0], validFunctions[1]), validFunctions[0]);
-			assert.notDeepEqual(enforce.func(validFunctions[0], validFunctions[1]), validFunctions[1]);
+			assert.deepEqual(enforce.function(validFunctions[0], validFunctions[1]), validFunctions[0]);
+			assert.notDeepEqual(enforce.function(validFunctions[0], validFunctions[1]), validFunctions[1]);
 		});
 
-		runNegativeTests('func', validFunctions[1]);
+		runNegativeTests('function', validFunctions[1]);
 	});
 });