+++ tests/methods/types/methodFunc.Test.js
@@ -3,7 +3,7 @@ import { functionData as data } from '../../TestUtil';
 import { testMethodType } from '../methodTestUtility';
 
 describe('method', () => {
-	describe('.func', () => {
+	describe('.function', () => {
 		testMethodType(assign({
 			coerce: false
 		}, data));