+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## Unreleased
-- Changed isBool to isBoolean, enforce.bool to enforce.boolean, and method.bool to method.Boolean
+- Changed isBool to isBoolean, enforce.bool to enforce.boolean, and method.bool to method.boolean
+- Changed isFunc to isFunction, enforce.func to enforce.function, and method.func to method.function
 
 ## 0.2.0 - 2019-01-18
 ### Added